# reticketify-backend-user/src/main/resources/openapi/openapi.yaml
openapi: 3.0.4
info:
  title: ReTicketify User - OpenAPI 3.0
  description: |-
    ReTicketify User
  version: 0.0.1-SNAPSHOT
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /login:
    post:
      tags:
        - login
      summary: Login User
      operationId: loginUser
      requestBody:
        description: Login User
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Login'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 10
        uuid:
          type: string
          example: 'f91a1345-a590-460f-ae55-475a8fb6bbdc'
        username:
          type: string
          example: theUser
        password:
          type: string
          example: '12345'
        name:
          type: string
          example: John
        cpf:
          type: string
          example: John
        email:
          type: string
          example: john@email.com
        phone:
          type: string
          example: '12345'
...
    Login:
      type: object
      properties:
        username:
          type: string
          example: theUser
        password:
          type: string
          example: '12345'